made speed proportional to health
-speed is now proportional to health
-made wind generator use gaussian random variable
-health and max health is now double
-boat now starts to slowly gain speed when sails out instead of gaining full speed

#story[1212]

+++ Commons/src/main/java/models/Competitor.java
@@ -75,11 +75,11 @@ public interface Competitor {
 
     void changeHeading(boolean upwind, double angle);
 
-    int getHealthLevel();
+    double getHealthLevel();
 
     void updateHealth(int delta);
 
-    int getMaxHealth();
+    double getMaxHealth();
 
     void setMaxHealth(int health);
